feat: dashboard improvements — loading flash fix, chart layout, goal widget
- Don't render default dashboard while user config is loading (prevents
  flash of default widgets and wasted data fetches)
- Move chart current value beside title (right-aligned) to reduce card height
- Add goal_progress widget type for showing goal progress bars on dashboard
- Extract shared GoalProgressBar component from Goals page for reuse
