+++ packages/api-spec/src/schemas/dashboard.ts
@@ -222,6 +222,17 @@ export const hrZonesConfigSchema = z
 
 export type HrZonesConfig = z.infer<typeof hrZonesConfigSchema>
 
+/**
+ * Goal progress widget - displays progress bars for user goals.
+ */
+export const goalProgressConfigSchema = z
+  .object({
+    compact: z.boolean().optional().meta({ description: 'Hide losing-tomorrow info (default: false)' }),
+  })
+  .meta({ id: 'GoalProgressConfig' })
+
+export type GoalProgressConfig = z.infer<typeof goalProgressConfigSchema>
+
 // =============================================================================
 // Widget Type Discriminated Union
 // =============================================================================
@@ -238,6 +249,7 @@ export const widgetTypeSchema = z.enum([
   'activity_summary',
   'quick_link',
   'hr_zones',
+  'goal_progress',
 ])
 
 export type WidgetType = z.infer<typeof widgetTypeSchema>
@@ -285,6 +297,11 @@ export const dashboardWidgetSchema = z.discriminatedUnion('type', [
     config: hrZonesConfigSchema,
     id: z.string().min(1).meta({ description: 'Unique widget ID' }),
     type: z.literal('hr_zones'),
+  }),
+  z.object({
+    config: goalProgressConfigSchema,
+    id: z.string().min(1).meta({ description: 'Unique widget ID' }),
+    type: z.literal('goal_progress'),
   }),
 ])
